Switch from native locale handling to QLocale
This ensures building on platforms where Qt is supported, rather than
just work on linux/gcc and msvc. Notably, mingw also works now.

This fixes issue 32 and 33.

Performance wise, QLocale is slightly slower, but for a 10mb json file
with a good mix of objects, arrays, doubles, integers and strings, the
timings was around 975 msec for using 'native locales' and 980 msec for
using QLocale. It was reproducibly approximately 5 msec slower for a
10mb file.

Note that the json_scanner.cc needs to be regenerated. I don't have the
same flex version, so the diff in that file was quite large.